restructure project, get basic version of tables working
