+++ requirements.txt
@@ -1,2 +1,3 @@
 beautifulsoup4==4.5.3
 requests==2.13.0
+terminaltables==3.1.0
